Split test_utils into smaller files
