+++ client/verta/tests/test_utils/test_utils.py
@@ -1,18 +1,9 @@
 # pylint: disable=unidiomatic-typecheck
 
-import six
-
-import pathlib2
-import subprocess
-import sys
-
-import verta
-from verta._internal_utils import _utils
-from verta._internal_utils import _file_utils
-from verta._internal_utils import _pip_requirements_utils
-from verta._internal_utils import pagination_utils
-
 import pytest
+import six
+from verta._internal_utils import _utils
+
 from .. import utils
 
 
@@ -261,118 +252,3 @@ class TestToBuiltin:
         builtin_str_list = _utils.to_builtin(str_list)
         assert type(builtin_str_list) is list
         assert all(type(val) is str for val in builtin_str_list)
-
-
-class TestPipRequirementsUtils:
-    def test_no_spacy_models_in_pip_freeze(self):
-        spacy = pytest.importorskip("spacy")
-        try:
-            spacy.load("en_core_web_sm")
-        except OSError:
-            pytest.skip("SpaCy en_core_web_sm model not installed")
-
-        # baseline: en_core_web_sm in pip freeze
-        assert list(
-            filter(
-                _pip_requirements_utils.SPACY_MODEL_REGEX.match,
-                six.ensure_str(
-                    subprocess.check_output([sys.executable, "-m", "pip", "freeze"]),
-                ).splitlines(),
-            )
-        )
-
-        # en_core_web_sm not in our pip freeze util
-        assert not list(
-            filter(
-                _pip_requirements_utils.SPACY_MODEL_REGEX.match,
-                _pip_requirements_utils.get_pip_freeze(),
-            )
-        )
-
-
-class TestFileUtils:
-    @pytest.mark.parametrize(
-        "input_filepath, expected_filepath",
-        [
-            ("data.csv", "data 1.csv"),
-            ("data 1.csv", "data 2.csv"),
-            ("my data.csv", "my data 1.csv"),
-            ("my data 1.csv", "my data 2.csv"),
-            ("archive.tar.gz", "archive.tar 1.gz"),
-            ("archive.tar 1.gz", "archive.tar 2.gz"),
-            ("my archive.tar.gz", "my archive.tar 1.gz"),
-            ("my archive.tar 1.gz", "my archive.tar 2.gz"),
-        ],
-    )
-    def test_increment_path(self, input_filepath, expected_filepath):
-        assert expected_filepath == _file_utils.increment_path(input_filepath)
-
-    @pytest.mark.parametrize(
-        "path, prefix_dir, expected",
-        [
-            # simple removal cases
-            ("files/data.csv", "files", "data.csv"),
-            ("files/data/data.csv", "files/data", "data.csv"),
-            # simple no-change cases
-            ("files/data.csv", "foo", "files/data.csv"),
-            ("files/data.csv", "fil", "files/data.csv"),
-            ("files/data/data.csv", "files/data.csv", "files/data/data.csv"),
-            # edge cases
-            ("data.csv", "data.csv", "data.csv"),
-            # examples from comments in fn
-            ("data/census-train.csv", "data/census", "data/census-train.csv"),
-            ("data/census/train.csv", "data/census", "train.csv"),
-            # remove "s3://"
-            (
-                "s3://verta-starter/census-train.csv",
-                "s3:",
-                "verta-starter/census-train.csv",
-            ),
-            (
-                "s3://verta-starter/census-train.csv",
-                "s3:/",
-                "verta-starter/census-train.csv",
-            ),
-            (
-                "s3://verta-starter/census-train.csv",
-                "s3://",
-                "verta-starter/census-train.csv",
-            ),
-        ],
-    )
-    def test_remove_prefix_dir(self, path, prefix_dir, expected):
-        assert _file_utils.remove_prefix_dir(path, prefix_dir) == expected
-
-    def test_flatten_file_trees(self, in_tempdir):
-        filepaths = {
-            "README.md",
-            "data/train.csv",
-            "data/test.csv",
-            "script.py",
-            "utils/data/clean.py",
-            "utils/misc/misc.py",
-        }
-        paths = ["README.md", "data", "script.py", "utils"]
-
-        # create files
-        for filepath in filepaths:
-            filepath = pathlib2.Path(filepath)
-            filepath.parent.mkdir(parents=True, exist_ok=True)
-            filepath.touch()
-
-        assert _file_utils.flatten_file_trees(paths) == filepaths
-
-
-class TestPaginationUtils:
-    def test_from_proto(self):
-        fn = pagination_utils.page_limit_from_proto
-        assert fn(-1) is None
-        assert fn(0) == 0
-        assert fn(1) == 1
-
-    def test_to_proto(self):
-        fn = pagination_utils.page_limit_to_proto
-        assert fn(None) == -1
-        assert fn(-1) == -1
-        assert fn(0) == 0
-        assert fn(1) == 1
